Added more REST logic, update product, display products and repaired code to be more in pep8 standards

+++ models/product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Column, Float, ForeignKey
+from sqlalchemy import Integer, String, Column, Float, ForeignKey, NUMERIC
 from database import Base
 
 
@@ -8,6 +8,6 @@ class Product(Base):
     p_id = Column(Integer, index=True, primary_key=True)
     p_name = Column(String(255), index=True)
     category = Column(String(100), index=True)
-    price = Column(Integer)
+    price = Column(NUMERIC)
     quantity = Column(Float)
     created_by = Column(Integer, ForeignKey('users.id'), nullable=True)
